+++ json_annotation/lib/src/allowed_keys_helpers.dart
@@ -2,11 +2,22 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+/// **DEPRECATED** Helper function used in generated code when
+/// `JsonSerializable.disallowUnrecognizedKeys` is `true`.
+///
+/// Should not be used directly.
+@Deprecated('Code generated with the latest `json_serializable` will use '
+    '`\$checkKeys` instead. This function will be removed in the next major '
+    'release.')
+void $checkAllowedKeys(Map map, Iterable<String> allowedKeys) {
+  $checkKeys(map, allowedKeys: allowedKeys?.toList());
+}
+
 /// Helper function used in generated code when
 /// `JsonSerializable.disallowUnrecognizedKeys` is `true`.
 ///
 /// Should not be used directly.
-void $checkAllowedKeys(Map map, Iterable<String> allowedKeys) {
+void $checkKeys(Map map, {Iterable<String> allowedKeys}) {
   if (map == null) return;
   var invalidKeys = map.keys.where((k) => !allowedKeys.contains(k));
   if (invalidKeys.isNotEmpty) {